+++ src/php/Compiler/CompilerFactory.php
@@ -11,6 +11,7 @@ use Phel\Compiler\Emitter\OutputEmitter\Munge;
 use Phel\Compiler\Emitter\OutputEmitter\NodeEmitterFactory;
 use Phel\Compiler\Emitter\OutputEmitter\SourceMap\SourceMapGenerator;
 use Phel\Compiler\Emitter\OutputEmitterInterface;
+use Phel\Compiler\Parser\QuasiquoteTransformer;
 
 final class CompilerFactory implements CompilerFactoryInterface
 {
