+++ src/php/Compiler/Parser/ExpressionParser/MetaParser.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel\Compiler\Parser;
+namespace Phel\Compiler\Parser\ExpressionParser;
 
 use Generator;
 use Phel\Compiler\Parser;
