+++ src/php/Compiler/Reader.php
@@ -11,8 +11,9 @@ use Phel\Compiler\Parser\ParserNode\NodeInterface;
 use Phel\Compiler\Parser\ParserNode\QuoteNode;
 use Phel\Compiler\Parser\ParserNode\SymbolNode;
 use Phel\Compiler\Parser\ParserNode\TriviaNodeInterface;
-use Phel\Compiler\ReadModel\CodeSnippet;
-use Phel\Compiler\ReadModel\ReaderResult;
+use Phel\Compiler\Parser\QuasiquoteTransformerInterface;
+use Phel\Compiler\Parser\ReadModel\CodeSnippet;
+use Phel\Compiler\Parser\ReadModel\ReaderResult;
 use Phel\Exceptions\ReaderException;
 use Phel\Lang\AbstractType;
 use Phel\Lang\IMeta;
