+++ src/php/Exceptions/ExceptionPrinterInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Phel\Exceptions;
 
-use Phel\Compiler\ReadModel\CodeSnippet;
+use Phel\Compiler\Parser\ReadModel\CodeSnippet;
 use Throwable;
 
 interface ExceptionPrinterInterface
