+++ src/php/Exceptions/ParserException.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Phel\Exceptions;
 
 use Exception;
-use Phel\Compiler\ReadModel\CodeSnippet;
+use Phel\Compiler\Parser\ReadModel\CodeSnippet;
 use Phel\Lang\SourceLocation;
 
 final class ParserException extends PhelCodeException
