+++ src/php/Exceptions/TextExceptionPrinter.php
@@ -7,7 +7,7 @@ namespace Phel\Exceptions;
 use Phel\Command\Repl\ColorStyle;
 use Phel\Compiler\Emitter\OutputEmitter\Munge;
 use Phel\Compiler\Emitter\OutputEmitter\SourceMap\SourceMapConsumer;
-use Phel\Compiler\ReadModel\CodeSnippet;
+use Phel\Compiler\Parser\ReadModel\CodeSnippet;
 use Phel\Lang\IFn;
 use Phel\Printer\Printer;
 use ReflectionClass;
